(FEATURE) - 8-game-ball-controller (#26)
* (FEATURE) - 8-game-ball-controller

https://github.com/kevinboogaard/Sports_World-Ball_Pit/issues/16

Allows the user to swap balls on the field.

* (ENHANCEMENT) - 8-game-ball-controller

https://github.com/kevinboogaard/Sports_World-Ball_Pit/issues/16

I've made better code by making the controller look for the neighbours- instead of looping through ALL the tiles in the grid.

* (FEATURE) - 8-game-ball-controller

https://github.com/kevinboogaard/Sports_World-Ball_Pit/issues/16

Added the feature to remove balls. Got a new idea where code runs with events so I have to enhance my old code now.

* (FEATURE) - 8-game-ball-controller

https://github.com/kevinboogaard/Sports_World-Ball_Pit/issues/16

Refactored so that the code works with events instead of calling
functions after functions.

* (FEATURE) - 8-game-ball-controller

https://github.com/kevinboogaard/Sports_World-Ball_Pit/issues/16

Added the functionality where the balls drop when a swap is done.

* (FEATURE) - 8-game-ball-controller

https://github.com/kevinboogaard/Sports_World-Ball_Pit/issues/16

Spawn balls when balls are removed!